+++ internal/model/user.go
@@ -7,7 +7,7 @@ import (
 )
 
 type User struct {
-	ID        uint      `json:"id" gorm:"primaryKey"`
+	ID        int64     `json:"id" gorm:"primaryKey"`
 	Username  string    `json:"username" gorm:"unique;not null"`
 	Password  string    `json:"-" gorm:"not null"`
 	CreatedAt time.Time `json:"created_at"`
